+++ src/components/drawer/drawer.styles.ts
@@ -92,9 +92,11 @@ export default css`
 
   .drawer__title {
     flex: 1 1 auto;
+    font: inherit;
     font-size: var(--sl-font-size-large);
     line-height: var(--sl-line-height-dense);
     padding: var(--header-spacing);
+    margin: 0;
   }
 
   .drawer__close {